src/nia/ui/requirements.txt
# Core dependencies
gradio>=4.0.0          # UI framework
anthropic>=0.7.0       # Claude API client
neo4j>=5.0.0          # Graph database
networkx>=3.0         # Graph visualization
matplotlib>=3.0.0     # Plotting
pyautogui>=0.9.54    # Computer control
pillow>=10.0.0       # Image handling
screeninfo>=0.8.1    # Multi-monitor support

# Remote access dependencies
ngrok>=0.12.0        # Port forwarding
pyngrok>=7.0.0       # ngrok Python wrapper
user-agents>=2.0     # Mobile detection

# Visualization dependencies
plotly>=5.0.0        # Interactive graphs
pywebview>=4.0       # Desktop window management

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=22.0.0
ruff>=0.1.0
mypy>=1.0.0

# Platform-specific dependencies
# macOS
pyobjc-core>=10.0; sys_platform == 'darwin'      # macOS API
pyobjc-framework-Quartz>=10.0; sys_platform == 'darwin'  # Screen capture on macOS
